UI Manager and Button UI
Created a UI manager and a Button UI to test the UI Manager Process function.
- Added Button hovering feedback when the Mouse hovers over the Button.

+++ SuperStrategy2000/SuperStrategy2000/Level.cpp
@@ -86,7 +86,7 @@ void Level::process() {
 				// Show the Range
 				this->m_levelGameBoard->breadthFirstSearch(
 					this->m_levelGameBoard->m_selectedSquare,
-					unit->getUnitSpeed(),
+					static_cast<int>(unit->getUnitSpeed()),
 					true
 				);
 			}
